--- src/test/groovy/modules/ManualsMenuModule.groovy
@@ -1,18 +1,0 @@
-package modules
-
-import geb.Module
-
-class ManualsMenuModule extends Module {
-    static content = {
-        toggle { $("div.menu a.manuals") }
-        linksContainer { $("#manuals-menu") }
-        links { linksContainer.find("a") }
-    }
-
-    void open() {
-        toggle.click()
-        waitFor { linksContainer.hasClass("animating") }
-    }
-
-
-}
